refactor: filter 클래스로 분리
