+++ src/App.js
@@ -1,6 +1,7 @@
 import CreateTodo from './components/CreateTodo.js'
 import DeleteTodo from './components/DeleteTodo.js'
 import EditTodo from './components/EditTodo.js'
+import FilterTodo from './components/FilterTodo.js'
 
 class App {
   constructor () {
@@ -8,11 +9,10 @@ class App {
     this.$todoList = document.querySelector('#todo-list')
     this.$filters = document.querySelector('.filters')
 
-    this.$filters.addEventListener('click', this.filterTodo)
-
     this.createTodo = new CreateTodo(this.$newTodo, this.loadTodo)
     this.deleteTodo = new DeleteTodo(this.$todoList, this.loadTodo)
     this.editTodo = new EditTodo(this.$todoList, this.loadTodo)
+    this.filterTodo = new FilterTodo(this.$filters, this.updateTodo)
 
     this.loadTodo()
   }
@@ -25,30 +25,9 @@ class App {
   updateTodo = (todos) => {
     this.$todoList.innerHTML = ''
     todos.map(
-      todo => this.$todoList.insertAdjacentHTML('beforeend',
-        this.todoItemTemplate(todo)),
+      todo => this.$todoList.insertAdjacentHTML('beforeend', this.todoItemTemplate(todo))
     )
     this.updateTodoCount(todos)
-  }
-
-  filterTodo = (e) => {
-    this.$filters.querySelector('.selected').classList.remove('selected')
-    e.target.classList.add('selected')
-
-    if (e.target.classList.contains('all')) {
-      this.updateTodo(this.todos)
-      this.updateTodoCount(this.todos)
-    }
-    if (e.target.classList.contains('active')) {
-      const todos = this.todos.filter(todo => !todo.completed)
-      this.updateTodo(todos)
-      this.updateTodoCount(todos)
-    }
-    if (e.target.classList.contains('completed')) {
-      const todos = this.todos.filter(todo => todo.completed)
-      this.updateTodo(todos)
-      this.updateTodoCount(todos)
-    }
   }
 
   todoItemTemplate = (todo) => {
